Incorporate changes from recent PR.
